fix/refactor: TextSpamClassifier and UserSpamStatusProcessor
fix: fixed KeyError bug in TextSpamClassifier
refactor:
- create new file 'spam_processor.py' for UserSpamStatusProcessor.
- change name from dataset.csv to spam_detaset.csv
